Mains securizations are done now, still dont the scheduler in place

+++ srcs/time_utils.c
@@ -6,7 +6,7 @@
 /*   By: larchimb <larchimb@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/08/12 09:22:10 by larchimb          #+#    #+#             */
-/*   Updated: 2026/08/14 16:50:30 by larchimb         ###   ########.fr       */
+/*   Updated: 2026/08/15 11:15:35 by larchimb         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ long long	get_exec_time(t_data *data)
 	return (actual_time - data->starting_time);
 }
 
-void	add_time_ms(struct timespec *ts, long time_to_add)
+void	add_time_ms(struct timespec *ts, long long time_to_add)
 {
 	clock_gettime(CLOCK_REALTIME, ts);
 	ts->tv_nsec += time_to_add;
